rbd-wnbd: improve image map error message
"rbd-wnbd map" commands are delegated to the centralized
ceph-rbd Windows service, which in turn will create the
actual image mapping. This allows the daemons to continue
running even after the current session ends.

rbd-wnbd communicates with the service through a named pipe.
If the operation fails, we're only getting an error code.

This change will print the according error message based on the
error code, suggesting the user to check the logs or use
foreground mode ("-f") for additional information.

We could also capture the entire subprocess console output and
print it in case of errors. This can be quite verbose, so we'll
keep it simple for now.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>
(cherry picked from commit 5bbcfc82970bd3a76536a7dd554096bec70ddc2d)