rgw: is_olh() predicate should use user.rgw.olh.ver xattr
The predicate which determines whether an object is an OLH object was previously based off of the
user.rgw.olh.info xattr, but this xattr is not set when the OLH object is first created. This
means that in cases where an error prevents that xattr from being set later, the OLH object is
incorrectly interpreted as an empty unversioned object. This change modifies the is_olh function
to use the user.rgw.olh.ver xattr. This xattr is set at the time of OLH creation and therefore
ensures consistency.

Fixes: https://tracker.ceph.com/issues/61359
Signed-off-by: Cory Snyder <csnyder@1111systems.com>