librbd: propagate blkdiscards in mirroring scenarios
Blkdiscards were not being propagated to the non-primary rbd mirror image
because ObjectCopyRequest::compute_zero_ops skipped the the zeroed extents
in a snapshot when not done as part of an rbd deep copy operation.
This fix updates the m_snapshot_sparse_bufferlist with the zeroed object extents
so a truncate op is sent to the non-primary image.

Fixes: https://tracker.ceph.com/issues/58852

Signed-off-by: Nithya B <nibalach@li-276caf4c-21dc-11b2-a85c-bc63e101829d.ibm.com>