rgw/sync: add copying of mdlog entries in fullsync()
In metadata incremental sync, RGWCloneMetaLogCoroutine copies mdlog entries
from the primary zone, so that it can serve those entries to other zones in
the case of failover (and it's promoted to the 'metadata master'), these entries
are not copied in metadata full sync. so a new zone that syncs all of its metadata
from full sync won't have any of the mdlog entries. if that new zone is promoted
to the metadata master, it won't be able to serve these entries to other zones
hat are behind in incremental sync. so those zones will think they're caught up,
but may be missing lots of metadata

Signed-off-by: Jiffin Tony Thottan <thottanjiffin@gmail.com>