client: ignore the MAY_READ access mode check if __FMODE_EXEC is set
The execlp(),etc system call will ignore the 'r' permission in VFS,
and allowed a file without 'r' permission still could be excuted if
'x' is set.

The kernel VFS and FUSE will pass the 'O_READONLY | __FMODE_EXEC' to
libcephfs in the open_flags when opening the file and if the
'fuse_default_permissions' option is disabled we can check the
'__FMODE_EXEC' flags. If '__FMODE_EXEC' is set then we should ignore
the 'r' permissio check.

Fixes: https://tracker.ceph.com/issues/61501
Signed-off-by: Xiubo Li <xiubli@redhat.com>