mon/Monitor: during shutdown don't accept new authentication and create new sessions During shutdown, the monitor is designed not to accept new authentication requests or create new sessions. However, a problem arises when the monitor marks its status as "shutdown" but still accepts new authentication requests and creates new sessions. This issue causes the monitor to fail when checking the session list.
To fix this problem, an update is implemented. With this fix,
a monitor in the "shutdown" state will correctly reject new authentication requests
and prevent the creation of new sessions.
This ensures that the monitor operates as intended during the shutdown process.

Fixes: https://tracker.ceph.com/issues/56192
Signed-off-by: Nitzan Mordechai <nmordech@redhat.com>
(cherry picked from commit 4a8c21e3763dad41555765143f4b76f3a543b4f4)