client: only when trying to clear suid/sgid will allow perm check pass
When an unprivileged user tries to change the file mode, which the
corresponding mode is already there, for example:

$ ll
-rw-r--r-- 1 root   root       0 Jun  1 12:36 fileA
$ su ${unprivileged-user} -c "chmod a+r fileA"

Currently it will succeed. While this is incorrect and won't comply
with the Posix and the existing filesystems.

We should fail it anyway and only when it's trying to clear the
suid/sgid will pass the permission check.

Introduced-by db10cfc8077("client: allow unprivileged users to clear suid/sgid")
Fixes: https://tracker.ceph.com/issues/61552
Signed-off-by: Xiubo Li <xiubli@redhat.com>