crimson/net: implement logic to move a io-handler to a new sid
Note that it is inevitable that the user can mark down the connection
while the protocol is trying to move the connection to another core.

In that case, the implementation should tolerate the racing, which
finally needs to cleanup resources correctly and dispatch reasonable
events.

Signed-off-by: Yingxin Cheng <yingxin.cheng@intel.com>
(cherry picked from commit b10f0bd3ccc919bfbe4d063bebabe651faf60812)