mgr: heartbeat_webhook module
This adds a new mgr module called heartbeat_webhook. This will
send webhooks to the specified url as long as the cluster status
is `HEALTH_OK`, and will stop sending them if the cluster enters
a warning or critical state. This allows heartbeat based
monitoring systems to easily and robustly alert for failed Ceph
clusters.

Signed-off-by: Grayson Head <grayson@graysonhead.net>