rgw_admin.cc: Handle pg_ver and source_zone explicitly
At present, the pg_ver and source_zone attrs are handled as if they're
strings, but are in fact unsigned integer types.  This means that when
dumped they yield invalid UTF-8 characters.  As well as hiding the
meaning of these attributes, this also means that we're not compliant
with RFC 8259 for JSON, which specifies that externally visible JSON
must be UTF-8 encoded [0].

[0]: https://rfc-editor.org/rfc/rfc8259#section-8.1

Signed-off-by: Tom Coldrick <tcoldrick@bloomberg.com>