mgr: load python module classes via a function call
Previously, the mgr would load a python module and then search through
the top-level items in the module for subclasses of either MgrModule or
MgrStandbyModule (for optional standby modules). The naming is slightly
confusing to a Python programmer but what are called modules in the mgr
are in fact classes.

This patch adds a new way of loading these module classes. The code will
call functions named "_get_ceph_mgr_module_class" and
"_get_ceph_mgr_standby_module_class" respectively. If these functions
are not found then the previous method is used. The function names are a
bit verbose but try to reflect that the mgr's modules are python
classes.

The benefits of this new method include:
* No need to search through a modules namespace, making it very clear
  and explicit what class will be used no chance of hitting a "only one
  MgrModule class is loaded from each plugin" error
* A small potential performance increase for the above when loading a
  large python module
* The ability to defer imports so that only the mgr needs so that the
  __init__.py doesn't have to unconditionally include (import) MgrModule
  classes. This can avoid "namespace pollution" when one python module
  under src/pybind/mgr imports another
* The ability to dynamically construct the classes, for example a
  factory function that differs mostly on naming/branding or other
  small ways be used for >1 mgr module.
* In time we may even be able to drop the necessity of using a specific
  subclass and instead by able to do more pythonic duck-typing (or
  typing.Protocol interfaces).

This work was mainly inspired by the 3rd bullet, specifically to
begin fixing: https://tracker.ceph.com/issues/61186

While I think this new method is generally better than the old I do not
plan on dropping the older method soon as to avoid having to change all
the modules quickly and I'm not clear on how many mgr modules in python
exist outside of the ceph git tree. However, I'd be in favor of
eventually dropping the older method in time.

Signed-off-by: John Mulligan <jmulligan@redhat.com>