Since linux kernel 5.9 (crypt version 1.22 and higher), cryptsetup 2.3.4 and higher, dm-crypt has support for not using read and write workqueues when performing encryption. These workqueues were invented to improve performance for spinning disks. Cloudflare engineers [1] figured out this had the opposite effect on flash media (Solid State Disks). To have I/O go directly to the device instead of going through a queue, options to bypass this behaviour in dm-crypt, have been created.
The changes in this commit will automatically disable this queueing for read and
write I/O when non-rotational media (flash) devices are opened by cryptsetup
when called in ceph-volume.

Performance tests performed by the author confirm that performance is improved.
Tests performed with FIO show that there is an improvement of 32% in
single thread 4K random write performance. A 3% improvement was observed for
4K random read performance.
But, maybe more importantly, the standard deviation in latency dropped by 71%
for writes resp. 6% for reads.

Above tests were performed on SAMSUNG MZQLB3T8HALS-00007 (FW Rev: EDA5702Q)
drive.

[1]: https://blog.cloudflare.com/speeding-up-linux-disk-encryption/

Signed-off-by: Stefan Kooman <stefan@kooman.org>