cephadm: run tcmu-runner through script to do restart on failure
Currently, cephadm runs tcmu-runner as a background
process inside the unit file deployed for iscsi
(rbd-target-api is the primary process). This means
if tcmu-runner crashes for whatever reason, systemd
will not attempt to restart it. This commits sets
up a script to serve as the container entrypoint
for the tcmu-runner container that will run
tcmu-runner and also restart it on failure
(unless there are too many failures in a short
period, at which point it gives up).

Fixes: https://tracker.ceph.com/issues/61667

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 1b76ea037647e22470ed069dc04feecb245ff0b5)