rgw: apply_olh_log should clear_olh before trimming olh log
When pending_removal is true and the log is trimmed prior to
clearing olh, it allows for the possibility that the olh is not
cleared (e.g. due to new pending xattrs on the olh obj), but the
olh log entries are removed.

Signed-off-by: Cory Snyder <csnyder@1111systems.com>
(cherry picked from commit dfbe4f9cb078bad9a0244ce433a9728f8bd6b912)

Conflicts:
	src/rgw/driver/rados/rgw_rados.cc

Cherry-pick notes:
- conflicts due to rgw_rados.cc being moved to new directory after Pacific