osd/Session: rename last_sent_epoch to projected_epoch
We always update last_sent_epoch to reflect the highest map
we think the peer will have, whether because we are sending
the incremental or due to the map_epoch on the MOSDPing.

last_epoch_sent was a confusing name since we also update it
when an MOSDPing from the peer reflects a higher epoch.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>