doc/radosgw: remove pipes from s3select.rst
Remove strange formatting introduced in
61d6813726fc08980903c6a6c50b0a5797ba7193.

Note to the documentation team: This is not a line-edit. This commit
includes nothing but the removal of pipes added to the left of much of
the text in this file. Several future commits will line-edit this file
and correct its formatting.

Signed-off-by: Zac Dover <zac.dover@proton.me>
(cherry picked from commit a4d302ee1a6e2f71263a898109ebed28e7e8eaca)