blk: don't load the VDO extblkdev plugin by default
It has been observed that, after moving the VDO support into
a separated plugin, there is a problem with loading this plugin
on some platforms. The patch deals with that in a drastical way:
by simply emptying the default list of preloaded extblkdev
plugins (`osd_extblkdev_plugins`). I doubt this is the-solution
-of-choice but I prefer to have a last-resort commit ready.

Fixes: https://tracker.ceph.com/issues/61725
Signed-off-by: Radosław Zarzyński <rzarzyns@redhat.com>