crimson/os/seastore/zbd: increase padding byte buffer size.
Before closing a segment, seastore writes a tail information
at the end of the segment, since ZBD zones are serially write
required, if a zone is not fully written to the end, it writes
padding bytes till the tail section.
Increase the size of write chunks to reduce the number of writes.

Signed-off-by: Aravind Ramesh <aravind.ramesh@wdc.com>