rgw/d3n: implement non-yielding d3n cache op
adds a simple `cache_read()` overload that uses a `on_notify()` callback
to issue the aio completion. this is similar to the yielding version,
except that it doesn't run on an executor. libaio will invoke this
callback directly, and we'll rely on BlockingAioThrottle to handle the
thread-safety

add a `struct state` to keep track of the `aiocb`. as with the librados
implementation, we use placement new to store this in `AioResult::user_data`

instead of storing a `bufferlist result` separately, preallocate the
buffer directly into `AioResult::data`; in case of errors, this buffer
is cleared before returning

Signed-off-by: Casey Bodley <cbodley@redhat.com>